Maze Generation complete. Read Description
Maze gen complete, but spawning the character is not. I need to get when the maze is done, to spawn the player at 0,0 on the maze (this is also where they respawn).
I also fixed up the character a bit so that you can almost not look through the walls (almost meaning barely and only on the edges).